refactor: move OnnxCrossEncoder to shared core module
Moves the ONNX cross-encoder reranker from rag/_retriever.py to a new
core/reranker.py module. This allows the memory module to use the
reranker without triggering rag's dependency check for markitdown.

Fixes the issue where `uvx --from "agent-cli[memory]"` would fail
because importing memory's retrieval code would trigger rag/__init__.py
which checks for all rag dependencies including markitdown.
