+++ agent_cli/memory/engine.py
@@ -25,8 +25,8 @@ if TYPE_CHECKING:
 
     from chromadb import Collection
 
+    from agent_cli.core.reranker import OnnxCrossEncoder
     from agent_cli.memory.models import ChatRequest
-    from agent_cli.rag._retriever import OnnxCrossEncoder
 
 LOGGER = logging.getLogger(__name__)
